language_bot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
Flask==1.1.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
multidict==4.7.5
nltk==3.5
python-dotenv==0.13.0
regex==2020.4.4
tqdm==4.46.0
websockets==8.1
Werkzeug==1.0.0
yarl==1.4.2
